track list: refactoring - split lines at 180 meridian outside exporters
