+++ src/lib/leaflet.control.track-list/lib/geo_file_exporters.js
@@ -1,59 +1,6 @@
-import L from 'leaflet';
 import utf8 from 'utf8';
 import escapeHtml from 'escape-html';
 import {saveNktk} from './parsers/nktk';
-
-function getSegmentLatForLng(latLng1, latLng2, lng) {
-    const deltaLat = latLng2.lat - latLng1.lat;
-    const deltaLng = latLng2.lng - latLng1.lng;
-    return latLng1.lat + deltaLat / deltaLng * (lng - latLng1.lng);
-}
-
-function splitLineAt180Meridian(latLngs) {
-    const wrappedLatLngs = latLngs.map((ll) => ll.wrap());
-    const newLines = [];
-    if (latLngs.length < 2) {
-        return newLines;
-    }
-
-    let newLine = [wrappedLatLngs[0]];
-    newLines.push(newLine);
-    for (let i = 1; i < wrappedLatLngs.length; i++) {
-        let latLng = wrappedLatLngs[i];
-        let prevLatLng = wrappedLatLngs[i - 1];
-        if (Math.abs(latLng.lng - prevLatLng.lng) <= 180) {
-            newLine.push(latLng);
-        } else {
-            let positiveLng = L.Util.wrapNum(latLng.lng, [0, 360]);
-            let positivePrevLng = L.Util.wrapNum(prevLatLng.lng, [0, 360]);
-            let splitLng = 180 - 0.000001 * Math.sign(latLng.lng);
-            let splitPrevLng = 180 - 0.000001 * Math.sign(prevLatLng.lng);
-            let splitLat = getSegmentLatForLng(
-                L.latLng(prevLatLng.lat, positivePrevLng),
-                L.latLng(latLng.lat, positiveLng),
-                splitLng
-            );
-            let splitPrevLat = getSegmentLatForLng(
-                L.latLng(prevLatLng.lat, positivePrevLng),
-                L.latLng(latLng.lat, positiveLng),
-                splitPrevLng
-            );
-            newLine.push(L.latLng(splitPrevLat, splitPrevLng).wrap());
-            newLine = [L.latLng(splitLat, splitLng).wrap(), latLng];
-            newLines.push(newLine);
-        }
-    }
-    return newLines;
-}
-
-function normalizeLines(lines) {
-    return (lines || [])
-        .map((segment) => splitLineAt180Meridian(segment))
-        .reduce((acc, cur) => {
-            acc.push(...cur);
-            return acc;
-        }, []);
-}
 
 function saveGpx(segments, name, points, withElevations=false) {
     const gpx = [];
@@ -77,15 +24,14 @@ function saveGpx(segments, name, points, withElevations=false) {
             gpx.push('\t</wpt>');
         }
     );
-    const normalizedSegments = normalizeLines(segments);
-    if (normalizedSegments.length > 0) {
+    if (segments.length > 0) {
         name = name || 'Track';
         name = escapeHtml(name);
         name = utf8.encode(name);
         gpx.push('\t<trk>');
         gpx.push('\t\t<name>' + name + '</name>');
 
-        for (let segment of normalizedSegments) {
+        for (let segment of segments) {
             gpx.push('\t\t<trkseg>');
             for (let point of segment) {
                 let x = point.lng.toFixed(6);
@@ -114,8 +60,7 @@ function saveKml(segments, name, points) {
     kml.push('\t<Document>');
     kml.push(`\t\t<name>${name}</name>`);
 
-    const normalizedSegments = normalizeLines(segments);
-    for (let [i, segment] of normalizedSegments.entries()) {
+    for (let [i, segment] of segments.entries()) {
         kml.push('\t\t<Placemark>');
         kml.push(`\t\t\t<name>Line ${(i + 1)}</name>`);
         kml.push('\t\t\t<LineString>');
